Mini Browse

+++ src/recipe/store/RecipesReducer.ts
@@ -1,13 +1,16 @@
-import ReduxHelper from '../../common/store/ReduxHelper';
-import { Recipe, RecipeAction, RecipeActionTypes, RecipesAction, RecipesState, RECIPES_STORE, RECIPE_STORE } from './types';
+import ReduxHelper, { ACTION } from '../../common/store/ReduxHelper';
+import { Recipe, RecipeAction, RecipeActionTypes, RECIPE_STORE } from './RecipeTypes';
+import { RecipesAction, RecipesState, RECIPES_STORE } from './RecipesTypes';
 
 const defaultState: RecipesState = ReduxHelper.getMapReducerDefaultState<Recipe>(RECIPES_STORE);
 
 const recipes = (state = defaultState, action: RecipesAction): RecipesState => {
+  if (state.ident !== action.store) return ReduxHelper.caseDefaultReducer(state, action, defaultState);
+
   if (action.store === RECIPE_STORE) {
     const recipeAction = action as RecipeAction;
     switch (recipeAction.type) {
-      case RecipeActionTypes.RECIPE_LOAD:
+      case ACTION.GET_SUCCESS:
         return ReduxHelper.setMapItem(state, recipeAction.data.slug, recipeAction.data);
       case RecipeActionTypes.RECIPE_DELETE:
         return ReduxHelper.deleteMapItem(state, recipeAction.data);
